Add Post create

+++ app/Http/Controllers/PostController.php
@@ -4,9 +4,22 @@ namespace App\Http\Controllers;
 
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Str;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
+    /**
+     * Instantiate a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware("auth")->except("show");
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -24,7 +37,7 @@ class PostController extends Controller
      */
     public function create()
     {
-        //
+        return view("posts.create");
     }
 
     /**
@@ -35,7 +48,28 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
-        //
+        $request->validate([
+            "title" => "required|unique:posts|max:290",
+            "content" => "required",
+            "image" => "image|max:1999",
+        ]);
+
+        $path = $request->hasFile("image")
+            ? $request->file("image")->store("images", "public")
+            : "images/no_entry_image.jpg";
+
+        Post::create([
+            "title" => $request->title,
+            "slug" => Str::slug($request->title),
+            "content" => $request->content,
+            "image" => $path,
+            "user_id" => $request->user()->id,
+        ]);
+
+        return redirect(route("dashboard"))->with(
+            "status",
+            "Pomyślnie dodano wpis"
+        );
     }
 
     /**
@@ -59,7 +93,9 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
-        //
+        return view("posts.edit", [
+            "post" => $post,
+        ]);
     }
 
     /**
@@ -71,7 +107,9 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
-        //
+        if (!Gate::allows("post-owner", $post)) {
+            abort(403);
+        }
     }
 
     /**
@@ -82,7 +120,12 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        if (!Gate::allows("post-owner", $post)) {
+            abort(403);
+        }
+
         $post->delete();
+
         return redirect(route("dashboard"))->with(
             "status",
             "Pomyślnie usunięto wpis"
